backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==3.1.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
httpie==2.2.0
idna==2.10
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
waitress==1.4.4
whitenoise==5.2.0
